Split up SSDP integration into modules
