+++ homeassistant/components/ssdp/const.py
@@ -1,0 +1,7 @@
+"""Constants for the SSDP integration."""
+
+from __future__ import annotations
+
+DOMAIN = "ssdp"
+SSDP_SCANNER = "scanner"
+UPNP_SERVER = "server"
